Move test type service up
